Removed redundant type casts

+++ src/main/java/de/thomas_oster/liblasercut/FloatPowerSpeedFocusFrequencyProperty.java
@@ -123,19 +123,19 @@ public class FloatPowerSpeedFocusFrequencyProperty implements LaserProperty
   {
     if ("power".equals(name))
     {
-      return (Float) this.getPower();
+      return this.getPower();
     }
     else if ("speed".equals(name))
     {
-      return (Float) this.getSpeed();
+      return this.getSpeed();
     }
     else if ("focus".equals(name))
     {
-      return (Float) this.getFocus();
+      return this.getFocus();
     }
     else if ("frequency".equals(name))
     {
-      return (Integer) this.getFrequency();
+      return this.getFrequency();
     }
     return null;
   }
@@ -170,11 +170,11 @@ public class FloatPowerSpeedFocusFrequencyProperty implements LaserProperty
   {
   if ("power".equals(name))
     {
-      return (Float) 0f;
+      return 0f;
     }
     else if ("speed".equals(name))
     {
-      return (Float) 0f;
+      return 0f;
     }
     else if ("focus".equals(name))
     {
@@ -195,11 +195,11 @@ public class FloatPowerSpeedFocusFrequencyProperty implements LaserProperty
   {
     if ("power".equals(name))
     {
-      return (Float) 100f;
+      return 100f;
     }
     else if ("speed".equals(name))
     {
-      return (Float) 100f;
+      return 100f;
     }
     else if ("focus".equals(name))
     {
