+++ src/main/java/de/thomas_oster/liblasercut/FloatPowerSpeedFocusProperty.java
@@ -111,15 +111,15 @@ public class FloatPowerSpeedFocusProperty implements LaserProperty
   {
     if ("power".equals(name))
     {
-      return (Float) this.getPower();
+      return this.getPower();
     }
     else if ("speed".equals(name))
     {
-      return (Float) this.getSpeed();
+      return this.getSpeed();
     }
     else if ("focus".equals(name))
     {
-      return (Float) this.getFocus();
+      return this.getFocus();
     }
     return null;
   }
@@ -150,11 +150,11 @@ public class FloatPowerSpeedFocusProperty implements LaserProperty
   {
   if ("power".equals(name))
     {
-      return (Float) 0f;
+      return 0f;
     }
     else if ("speed".equals(name))
     {
-      return (Float) 0f;
+      return 0f;
     }
     else if ("focus".equals(name))
     {
@@ -175,11 +175,11 @@ public class FloatPowerSpeedFocusProperty implements LaserProperty
   {
     if ("power".equals(name))
     {
-      return (Float) 100f;
+      return 100f;
     }
     else if ("speed".equals(name))
     {
-      return (Float) 100f;
+      return 100f;
     }
     else if ("focus".equals(name))
     {
